backup-python/deploy/upload/pilotis/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
httpx>=0.26.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
reportlab>=4.0.0
a2wsgi>=1.10.0
